fixing double to int

+++ sum_array.cpp
@@ -12,12 +12,12 @@ sum_array(int N, int A[])
    // and return that sum to the caller
 
    // Put your code here, return the correct result
-   double sum = 0.0;
+   int sum = 0;
 
    for (int i; i < N; i++) {
       sum += A[i];
    }
-   
+
    return sum;
 }
